Refactor NamespacedPathnameBuilder
