+++ src/Twig/NamespacedPathnameBuilder.php
@@ -25,7 +25,7 @@
 
 declare(strict_types=1);
 
-namespace Machinateur\TwigBlockValidator\Service;
+namespace Machinateur\TwigBlockValidator\Twig;
 
 use Symfony\Component\Finder\SplFileInfo;
 use Twig\Error\LoaderError;
@@ -34,7 +34,7 @@ use Twig\Loader\FilesystemLoader;
 /**
  * This class is not available via dependency injection,
  *  as it requires a twig loader instance as constructor argument,
- *   which belongs to the respective twig-environment.
+ *   which belongs to the respective {@see BlockValidatorEnvironment}.
  */
 class NamespacedPathnameBuilder
 {
